requirements.txt
newspaper3k>=0.2.8
feedparser>=6.0.10
jinja2>=3.1.2
requests>=2.31.0
resend>=0.6.0
pyyaml>=6.0.1
beautifulsoup4>=4.12.2
python-dotenv>=1.0.0

RESEND_API_KEY=your_api_key_here
